Develop ai

See merge request s10-final/S10P31B103!152

+++ ai-server/app/routers/animation.py
@@ -38,11 +38,15 @@ class CharacterCreateRequest(BaseModel):
 
 
 @router.post("/create")
-async def create_joint_gif(request: CharacterCreateRequest):
+def crate_character(request: CharacterCreateRequest):
     character_id = request.character_id
     motion = request.motion_name
     s3_img_url = request.img_url
 
+    create_gif(character_id, motion, s3_img_url)
+
+
+async def create_gif(character_id, motion, s3_img_url):
     try:
         # 이미지 저장 경로
         IMG_DIR = "temp_image"
@@ -52,7 +56,7 @@ async def create_joint_gif(request: CharacterCreateRequest):
         is_downloaded = await get_img_s3(s3_img_url)
 
         if not is_downloaded:
-            return JSONResponse(content={"error": "Fast API ERROR : s3 이미지 다운로드 실패"}, status_code=500)
+            return JSONResponse(content={"error": "Fast API 에러 : s3 이미지 다운로드 실패"}, status_code=500)
 
         image_path = f"temp_image/{s3_img_url}"
 
@@ -80,20 +84,20 @@ async def create_joint_gif(request: CharacterCreateRequest):
         image_to_animation(image_path, gif_path, motion_cfg_fn, retarget_cfg_fn, character_id, motion)
 
         # gif 압축
-        # is_compressed = await gif_compress(
-        #     original_path=gif_path + f"/{character_id}_{motion}.gif",
-        #     compressed_path=gif_path + f"/{character_id}_{motion}_compressed.gif"
-        # )
-        #
-        # if not is_compressed:
-        #     print("압축 실패")
-        #     return JSONResponse(content={"error": "Fast API ERROR : s3 gif 업로드 실패"}, status_code=500)
+        is_compressed = await gif_compress(
+            original_path=gif_path + f"/{character_id}_{motion}.gif",
+            compressed_path=gif_path + f"/{character_id}_{motion}_compressed.gif"
+        )
+
+        if not is_compressed:
+            print("압축 실패")
+            return JSONResponse(content={"error": "Fast API 에러 : s3 gif 업로드 실패"}, status_code=500)
 
         # S3에 gif 업로드
         is_saved = await save_gif_s3(gif_path, character_id, motion)
 
         if not is_saved:
-            return JSONResponse(content={"error": "Fast API ERROR : s3 gif 업로드 실패"}, status_code=500)
+            return JSONResponse(content={"error": "Fast API 에러 : s3 gif 업로드 실패"}, status_code=500)
 
         # 로컬에 저장된 img, gif 삭제
         os.remove(image_path)
@@ -103,7 +107,7 @@ async def create_joint_gif(request: CharacterCreateRequest):
 
     except Exception as e:
         logger.error("create_joint_gif => 에러 발생", exc_info=True)
-        return JSONResponse(content={"error": f"Fast API ERROR : create_joint_gif => {e}"}, status_code=500)
+        return JSONResponse(content={"error": f"Fast API 에러 : create_joint_gif => {e}"}, status_code=500)
 
 
 # S3에서 img 불러오기
@@ -126,7 +130,7 @@ async def save_gif_s3(gif_path, character_id, motion):
     try:
         s3.upload_file(
             Bucket=os.getenv("S3_BUCKET"),
-            Filename=gif_path + f'/{character_id}_{motion}.gif',  # 업로드할 파일
+            Filename=gif_path + f'/{character_id}_{motion}_compressed.gif',  # 업로드할 파일
             Key=f'{os.getenv("S3_PATH")}/{character_id}_{motion}.gif',  # s3 저장 경로
             ExtraArgs={'ContentType': 'image/gif'}
         )
@@ -138,7 +142,7 @@ async def save_gif_s3(gif_path, character_id, motion):
 
 
 # gif 압축
-async def gif_compress(original_path, compressed_path):
+async def gif_compress(original_path, compressed_path, quality=0.5):
     try:
         with Image.open(original_path) as origin:
             # 프레임 목록 추출
@@ -146,7 +150,7 @@ async def gif_compress(original_path, compressed_path):
 
             # 각 프레임을 조정하여 크기를 줄임
             compressed_frames = [
-                frame.resize((int(frame.width * 0.7), int(frame.height * 0.7)), Image.Resampling.LANCZOS)
+                frame.resize((int(frame.width * quality), int(frame.height * quality)), Image.Resampling.LANCZOS)
                 for frame in frames
             ]
